requirements.txt
# Core dependencies
networkx>=3.0
python-dateutil>=2.8.0

# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
